requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
djangorestframework==3.13.1
gunicorn==20.1.0
isort==5.10.1
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.0.0
